+++ js/main.js
@@ -1,3 +1,4 @@
-import './get-similar-advertisments.js';
-import './create-similar-advertisments.js';
+import {getAdvertisment} from './get-advertisment.js';
+import {drawAdvertisments} from './draw-advertisment.js';
 
+drawAdvertisments(getAdvertisment(10));
